# vfs: add memFile.Flush, remove flushFile