Add `memFile.Flush` so that users of `memFile` (such as
`sstable.Writer`) know that it is buffered and don't wrap it in a
`bufio.Writer`. This prevents unnecessary buffering and allows the
`memFile` invariants to work better. Benchmarked that this was not a
performance problem with `sstable.BenchmarkWriter`. If anything, this
might be slightly faster:

```
name       old time/op    new time/op    delta
Writer-16    80.7ms ± 2%    80.1ms ± 1%    ~     (p=0.400 n=10+9)

name       old alloc/op   new alloc/op   delta
Writer-16    23.2MB ± 0%    23.1MB ± 0%  -0.02%  (p=0.000 n=10+9)

name       old allocs/op  new allocs/op  delta
Writer-16       709 ± 0%       707 ± 0%  -0.28%  (p=0.000 n=10+10)
```